+++ enums/fileMap.ts
@@ -1,3 +1,4 @@
 export enum FileMap {
 	BRAIN = "./data/brain.json",
+	LOGO = "/brainLogo.svg",
 }