+++ enums/siteMap.ts
@@ -1,6 +1,8 @@
 export enum SiteMap {
 	HOME = "/",
+	TALK_TO_A_BRAIN = "/talkToABrain",
 	CONVERSATION = "/conversation",
-	CREATE_BRAIN = "createBrain",
-
+	CREATE_BRAIN = "/createBrain",
+	FEED_BRAIN = "/feedBrain",
+	TUTORIAL = "/tutorial",
 }