+++ routes/_app.tsx
@@ -1,3 +1,4 @@
+import { FileMap } from "../enums/fileMap.ts";
 import { define } from "../utils.ts";
 
 export default define.page(function App({ Component }) {
@@ -6,10 +7,10 @@ export default define.page(function App({ Component }) {
       <head>
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-				<link rel="icon" href="/brainCropped.svg" type="image/svg+xml" />
+				<link rel="icon" href={FileMap.LOGO} type="image/svg+xml" />
         <title>Brain In A Jar</title>
       </head>
-      <body f-client-nav>
+      <body class="bg-bubble-gum-blue">
         <Component />
 			</body>
     </html>