+++ routes/talkToABrain.tsx
@@ -1,0 +1,21 @@
+import { Partial } from "fresh/runtime";
+import BrainLogo from "../components/BrainInAJarLogo.tsx";
+import { define } from "../utils.ts";
+import ChatForm from "../components/ChatForm.tsx";
+import { SiteMap } from "../enums/siteMap.ts";
+
+export default define.page(function TalkToABrain(_ctx){
+	return (
+		<div class="px-4 py-8 mx-auto h-fit">
+			<BrainLogo />
+			<div class="mx-auto flex flex-col items-center w-[65dvw] h-fit min-h-[90dvh] bg-black py-4">
+				<div class="mx-2 flex flex-col" f-client-nav>
+					<Partial name="chatResponse">{}</Partial>
+					<Partial name="chatResponseForm">
+						<ChatForm url={SiteMap.CONVERSATION} />
+					</Partial>
+				</div>
+			</div>
+		</div>
+	);
+});